Move Microbit specific components
