+++ apps/test/unit/lib/kits/maker/boards/microBit/AccelerometerTest.js
@@ -1,6 +1,6 @@
-import {expect} from '../../../../util/reconfiguredChai';
+import {expect} from '../../../../../../util/reconfiguredChai';
 import Accelerometer from '@cdo/apps/lib/kits/maker/boards/microBit/Accelerometer';
-import {MicrobitStubBoard} from './boards/makeStubBoard';
+import {MicrobitStubBoard} from '../makeStubBoard';
 import {sensor_channels} from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitConstants';
 import sinon from 'sinon';
